requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
torch>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
matplotlib>=3.7.0
tensorboard>=2.13.0
pytest>=7.4.0
pydantic>=2.0.0
rich>=13.0.0
